Aggiornamenti e nuove funzionalità per parcheggio
Questo commit introduce una serie di miglioramenti e nuove funzionalità all'applicazione web ASP.NET Core, ProgettoAPPWEB24. Tra le modifiche principali:

- Aggiunta del supporto per il login esterno e la conferma di registrazione via email, migliorando l'esperienza utente nell'autenticazione.
- Implementate modifiche significative alla struttura del database, inclusa l'aggiunta di `IdParcheggio` a varie entità e l'aggiornamento dei tipi di dati da `decimal` a `double` per i campi `Ricarica` e `Sosta`.
- Rimossi file di migrazione obsoleti, riflettendo l'annullamento o la rifacimento di specifiche migrazioni.
- Introdotta una nuova pagina Razor `CercaBiglietto` per la ricerca e il pagamento dei biglietti di parcheggio, oltre a una pagina `EditModel` per la modifica dei costi di parcheggio.
- Aggiornate le pagine esistenti e i modelli per allinearsi alle modifiche del modello di dati e migliorare l'interazione dell'utente.
- Rimossa la dipendenza da `IHttpContextAccessor` in `Ricarica.cshtml.cs` e `Sosta.cshtml.cs`, semplificando la gestione dei parcheggi e dei biglietti.
- Aggiunto un riferimento al pacchetto `Microsoft.EntityFrameworkCore.SqlServer`, indicando un'espansione nelle opzioni di database supportate.

Queste modifiche estendono la funzionalità dell'applicazione, migliorando la gestione dei dati di parcheggio e l'esperienza utente generale.

+++ ProgettoAPPWEB24/Models/Costo.cs
@@ -2,11 +2,11 @@
 
 namespace ProgettoAPPWEB24.Models
 {
-	public class Costo : IHasId
-	{
-		public int Id { get; set; }
-		public int IdParcheggio { get; set; }
-		public decimal Ricarica { get; set; }
-		public decimal Sosta { get; set; }
-	}
+    public class Costo : IHasId
+    {
+        public int Id { get; set; }
+        public int IdParcheggio { get; set; }
+        public double Ricarica { get; set; }
+        public double Sosta { get; set; }
+    }
 }
